Revision of hotfix for Issue #89 ensures alliance model exists before checking ID

+++ celerytask/tasks.py
@@ -432,9 +432,10 @@ def run_corp_update():
                 if all_corp_info.corporation_id != settings.CORP_ID:
                     if not all_corp_info.is_blue:
                         all_corp_info.delete()
-            elif all_corp_info.alliance.alliance_id != settings.ALLIANCE_ID:
-                if not all_corp_info.is_blue:
-                    all_corp_info.delete()
+            elif all_corp_info.alliance is not None:
+                if all_corp_info.alliance.alliance_id != settings.ALLIANCE_ID:
+                    if not all_corp_info.is_blue:
+                        all_corp_info.delete()
 
 
         # Check the alliances
